cors.go changes
